requirements.txt
streamlit>=1.22.0
pandas>=1.3.0
python-docx>=0.8.11
openpyxl>=3.0.9
odfpy>=1.4.1
xlrd>=2.0.1
pyxlsb>=1.0.9